Move around test files to avoid breaking builds
While renaming the `faq-databases` to `faqs` libs
